+++ gatsby-node.js
@@ -21,7 +21,7 @@ exports.onCreateNode = ({ node, getNode, actions }) => {
 
 exports.createPages = ({ graphql, actions }) => {
     const { createPage } = actions;
-    const blogPostTemplate = path.resolve("src/templates/blog-post.tsx");
+    const blogPostTemplate = path.resolve("src/templates/BlogPost.tsx");
     return graphql(`
         {
             allMarkdownRemark(
