+++ src/components/ThemeToggle.tsx
@@ -1,0 +1,6 @@
+import { useTheme } from "../theme";
+
+export const ThemeToggle: React.FC = () => {
+    const { toggleTheme } = useTheme();
+    return <button onClick={toggleTheme}>Toggle theme</button>;
+};
