+++ src/components/index.ts
@@ -1,29 +1,13 @@
-import { BlogContent } from "./blog-content";
-import { BlogHeader } from "./blog-header";
-import { BlogImage } from "./blog-image";
-import { Card } from "./card";
-import { Header } from "../layouts/header";
-import { Link } from "./link";
-import { MarkdownWrapper } from "./markdown-wrapper";
-import { NextPreviousLinks } from "./next-previous-links";
-import { PostImage } from "./post-image";
-import { PresentationGridLayout } from "./presentation-grid-layout";
-import { ProfileImage } from "./profile-image";
-import { ResumeContent } from "./resume-content";
-import { SEO } from "./seo";
-
-export {
-    BlogContent,
-    BlogHeader,
-    BlogImage,
-    Card,
-    Header,
-    Link,
-    MarkdownWrapper,
-    NextPreviousLinks,
-    PostImage,
-    PresentationGridLayout,
-    ProfileImage,
-    ResumeContent,
-    SEO,
-};
+export * from "./BlogContent";
+export * from "./BlogHeader";
+export * from "./BlogImage";
+export * from "./Card";
+export * from "../layouts/Header";
+export * from "./Link";
+export * from "./MarkdownWrapper";
+export * from "./NextPreviousLinks";
+export * from "./PostImage";
+export * from "./PresentationGridLayout";
+export * from "./ProfileImage";
+export * from "./ResumeContent";
+export * from "./Seo";
