--- src/components/theme-toggle.tsx
@@ -1,6 +1,0 @@
-import { useTheme } from "../theme";
-
-export const ThemeToggle: React.FC = () => {
-    const { toggleTheme } = useTheme();
-    return <button onClick={toggleTheme}>Toggle theme</button>;
-};
