+++ src/theme/index.ts
@@ -1,3 +1,3 @@
 export * from "./breakpoints";
 export * from "./colors";
-export * from "./theme-provider";
+export * from "./ThemeProvider";
